requirements.txt
Flask==3.1.1
gunicorn==21.2.0
openai==1.95.1
python-dotenv==1.1.1
PyMuPDF==1.22.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
certifi==2025.7.9
textblob==0.19.0
fpdf==1.7.2
idna==3.10
nltk==3.9.1
regex==2024.11.6
tqdm==4.67.1
typing_extensions==4.14.1
Flask-Talisman==1.0.0


